backend/requirements.txt
flask==3.0.0
Flask-CORS==5.0.0
requests==2.31.0
python-dotenv==1.0.0
waitress==2.1.2
langchain==0.3.14
langchain-community==0.3.14
langchain-ollama==0.2.2
langchain-core==0.3.29
beautifulsoup4==4.12.3
langgraph==0.2.60
langchain-postgres==0.0.12
psycopg-binary==3.2.3
psycopg2==2.9.10
duckduckgo-search==7.5.1